backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0
django-cors-headers>=3.14
whitenoise